Color, Typo, Link, Font-Type & Size
Changed the background color to be lighter but still dark. Fixed a few typos in the header and the links. Changed the font color to fit with the new background. Changed the fonts to times new roman and upped the size on p to 25px and on a to 17px